files refector, bug fix, refine doc

+++ auto_round/export_to_autogptq.py
@@ -44,6 +44,7 @@ from safetensors.torch import save_file as safe_save
 from os.path import join, isfile, isdir
 import copy
 import json
+from .utils import quant_weight_w_scale
 
 
 def compress_model(
@@ -95,7 +96,8 @@ def compress_model(
         scale_dtype = v["scale_dtype"]
         m = get_module(compressed_model, k)
         fp_weight = m.weight.data
-        scale = torch.tensor(v["scale"], dtype=scale_dtype)
+        # scale = torch.tensor(v["scale"], dtype=scale_dtype)
+        scale = torch.tensor(v["scale"], dtype=torch.float32)
         zp = None if scheme == "sym" else torch.tensor(v["zp"], dtype=torch.int32)
         int_weight = quant_weight_w_scale(fp_weight, scale, zp, group_size)
         int_weight = int_weight.type(torch.int32)
@@ -200,4 +202,3 @@ def save_quantized_to_autogptq(model, save_dir: str, bits=4, group_size=128, sym
     quantization_config.model_file_base_name = model_base_name
 
     quantization_config.save_pretrained(save_dir)
-
